Fixed about us spheres and font responsiveness
