+++ next/src/pages/MiniSupport.tsx
@@ -1,43 +1,68 @@
-import React from 'react'
+import React, { useState } from "react";
+import Image from "next/image";
+import Footer from "../components/footer/Footer";
+import Navbar from "../components/navbar/Navbar";
+import { NavbarOpenHandler, NavbarType } from "../components/navbar/types";
 import styled from 'styled-components'
-import Otter from '../svgs/otter.png'
-import Image from 'next/image';
 import Link from 'next/link'
 import { device } from '../styles/device'
 
-import { SectionFadeInFromLeft, SectionFadeInFromRight } from "../styles/motion"
 
-type Props = {}
+import Otter from '../svgs/otter.png'
 
-const Container = styled.div`
-  @media ${device.laptop} {
-    height: 100vh;
-  }
+const Text = styled.p`
+  color: white;
+  font-size: 3vmin;
+  text-align: center;
 `
 
-const HeadingContainer = styled.div`
+const Grid = styled.div`
   display: flex;
   justify-content: center;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 70vw;
+  max-width: 75vw;
+  border-radius: 0.5rem;
+  margin-bottom: 10vh;
+  background: radial-gradient(50% 50% at 50% 50%, rgba(146, 67, 166, 0.2407) 0%, rgba(119, 158, 237, 0.83) 100%);
+ 
 `
 
-const Heading = styled.div`
-  color: #A09FE3;
-  font-family: 'Raleway';
-  font-weight: 800;
-  font-size: 30px;
-  @media ${device.tablet} {
-    font-size: 3.5vw;
-  }
-`;
-
-const BodyContainer = styled.div`
+const SmallGrid = styled.div`
   display: flex;
+  justify-content: center;
   flex-direction: column;
-  padding: 10vh 0;
+  flex: 2;
   @media ${device.tablet} {
+    min-height: 100%;
+    flex-wrap: wrap;
     flex-direction: row;
-    padding: 10vh 30vw;
+    flex: 3;
   }
+`
+
+const OurSponsorsCol = styled.div`
+  flex: 2;
+  padding: 0vmin 2.5vmin;
+  
+`
+
+const SponsorCol = styled.div`
+  flex: 3;
+  display: flex; 
+  justify-content: center;
+  max-height: 100%;
+  max-width: 100%;
+  border-radius: 0.5rem;
+  padding: 1vmin;
+  @media ${device.tablet} {
+    flex-basis: calc(50% - 40px);  
+    flex-direction: column;
+    padding: 4vmin 3vmin;
+  }
+
+
 `
 
 const TextContainer = styled.div`
@@ -46,47 +71,26 @@ const TextContainer = styled.div`
   justify-content: center;
   align-items: center;
   height: 100%;
-  padding: 0 3vw;
-`
-
-const H3 = styled.div`
-  color: #9B9B9B;
-  font-family: 'Raleway';
-  font-weight: 800;
-  font-size: 20px;
-  @media ${device.tablet} {
-    font-size: 2.5vw;
-  }
-`;
-
-const Text = styled.p`
-  color: #9B9B9B;
 `
 
 const ButtonContainer = styled.div`
   display: flex;
 `
 
-const FlexCenter = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-const ImgContainer = styled.div`
-  width: 50vw;
+const H3 = styled.div`
+  color: white;
+  font-family: 'Raleway';
+  font-weight: 800;
+  font-size: 4vmin;
+`;
 
-  @media ${device.laptop} {
-    width: 350px;
-    height: 400px;
-  }
-`
 
 const Button = styled.button`
-  background-color:  #9B9B9B;
+  background-color: white;
   margin: 10px;
-  padding: 10px 45px;
-  font-size: 18px;
-  color: white;
+  padding: 1.5vmin 2.5vmin;
+  font-size: 2.5vmin;
+  color: #817fff;
   border: none;
   border-radius: 6px;
   cursor: pointer;
@@ -94,35 +98,100 @@ const Button = styled.button`
   &:hover {
     transform: scale(1.05);
   }
+  @media ${device.tablet} {
+    padding: 2vmin 3vmin;
+  }
 `
 
+const ImgContainer = styled.div`
 
-export default function Support({ }: Props) {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: 30vw;
+
+  @media ${device.laptop} {
+    width: 350px;
+    height: 400px;
+  }
+
+`
+
+const HeadingContainer = styled.div`
+  display: flex;
+  justify-content: center;
+`
+
+const Heading = styled.div`
+  color: var(--accent-darker-purple);
+  font-family: 'Raleway';
+  font-weight: 800;
+  font-size: 5vmin;
+`;
+
+
+
+export default function Support() {
   return (
-    <Container>
+    <div>
       <HeadingContainer>
         <Heading>Support CSESoc</Heading>
       </HeadingContainer>
-      <BodyContainer>
-        <SectionFadeInFromLeft>
+      <ImgContainer>
+        <Image src={Otter} />
+      </ImgContainer>
+
+      <Grid>
+
+        <OurSponsorsCol>
           <TextContainer>
             <H3>Our Sponsors</H3>
+            <Text>Check out our very cool sponsors</Text>
             <ButtonContainer>
               <Link href="/Sponsors">
-                <Button>view our sponsors</Button>
+                <Button>View our sponsors</Button>
               </Link>
             </ButtonContainer>
-            <Text>Check out our very cool sponsors</Text>
           </TextContainer>
-        </SectionFadeInFromLeft>
-        <SectionFadeInFromRight>
-          <FlexCenter>
-            <ImgContainer>
-              <Image src={Otter} />
-            </ImgContainer>
-          </FlexCenter>
-        </SectionFadeInFromRight>
-      </BodyContainer>
-    </Container>
-  )
+        </OurSponsorsCol>
+
+        <SmallGrid>
+          <SponsorCol style={{ backgroundColor: 'rgba(0, 71, 255, 0.33)' }}>
+            <Image
+              src={`/assets/sponsors_white/atl.svg`}
+              width={100}
+              height={30}
+              objectFit="contain"
+            />
+          </SponsorCol>
+          <SponsorCol style={{ backgroundColor: 'rgba(82, 130, 255, 0.47)' }}>
+            <Image
+              src={`/assets/sponsors_white/imc.svg`}
+              width={60}
+              height={50}
+              objectFit="contain"
+            />
+          </SponsorCol>
+
+          <SponsorCol style={{ backgroundColor: 'rgba(48, 93, 255, 0.2)' }}>
+            <Image
+              src={`/assets/sponsors_white/deloitte.svg`}
+              width={50}
+              height={40}
+              objectFit="contain"
+            />
+          </SponsorCol>
+          <SponsorCol style={{ backgroundColor: 'rgba(122, 137, 236, 0.27)' }}>
+            <Image
+              src={`/assets/sponsors_white/js.svg`}
+              width={50}
+              height={55}
+              objectFit="contain"
+            />
+          </SponsorCol>
+        </SmallGrid>
+      </Grid>
+    </div>
+  );
+
 }
